# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Habit Journal API
  description: API for managing habits in a habit journal application
  version: 1.0.0
  contact:
    name: Habit Journal Team
    url: https://github.com/habitjournal/habit-journal-api
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.habitjournal.com
    description: Production server
paths:
  /v1/habits:
    post:
      summary: Create a new habit
      description: Creates a new habit with the provided name
      operationId: createHabit
      tags:
        - Habits
      requestBody:
        description: Habit creation request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HabitRequestDTO'
            examples:
              sample-habit:
                summary: Sample habit request
                value:
                  name: "Exercise daily"
      responses:
        '201':
          description: Habit created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitResponseDTO'
              examples:
                sample-response:
                  summary: Sample created habit
                  value:
                    id: 1
                    name: "Exercise daily"
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all habits
      description: Retrieves a list of all habits
      operationId: getAllHabits
      tags:
        - Habits
      responses:
        '200':
          description: Successful retrieval of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HabitResponseDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/habits/{id}:
    get:
      summary: Get habit by ID
      description: Retrieves a specific habit by its ID
      operationId: getHabitById
      tags:
        - Habits
      parameters:
        - name: id
          in: path
          description: ID of the habit to retrieve
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Successful retrieval of habit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitResponseDTO'
        '404':
          description: Habit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update habit
      description: Updates an existing habit
      operationId: updateHabit
      tags:
        - Habits
      parameters:
        - name: id
          in: path
          description: ID of the habit to update
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        description: Habit update request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HabitRequestDTO'
      responses:
        '200':
          description: Habit updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitResponseDTO'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Habit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete habit
      description: Deletes a habit by its ID
      operationId: deleteHabit
      tags:
        - Habits
      parameters:
        - name: id
          in: path
          description: ID of the habit to delete
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Habit deleted successfully
        '404':
          description: Habit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    HabitRequestDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the habit
          minLength: 1
          maxLength: 100
          example: "Exercise daily"
    HabitResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the habit
          example: 1
        name:
          type: string
          description: Name of the habit
          example: "Exercise daily"
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the error occurred
        status:
          type: integer
          format: int32
          description: HTTP status code
        error:
          type: string
          description: HTTP status text
        message:
          type: string
          description: Detailed error message
        path:
          type: string
          description: The path where the error occurred
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []